src/data/usecases/add-account/db-add-account.ts
import { AccountModel, AddAccount, AddAccountModel, Hasher, AddAccountRepository } from './db-add-account-protocols'

export class DbAddAccount implements AddAccount {
  private readonly hasher: Hasher
  private readonly addAccountRepo: AddAccountRepository

  constructor (hasher: Hasher, addAccountRepo: AddAccountRepository) {
    this.hasher = hasher
    this.addAccountRepo = addAccountRepo
  }

  async add (accountData: AddAccountModel): Promise<AccountModel> {
    const hashedPassword = await this.hasher.hash(accountData.password)
    const account = await this.addAccountRepo.add(Object.assign({}, accountData, { password: hashedPassword }))
    return new Promise(resolve => resolve(account))
  }
}
